chore: move types to models.ts
