+++ apps/passport-server/src/database/models.ts
@@ -224,3 +224,23 @@ export interface ZuconnectTicketDB {
   is_deleted: boolean;
   is_mock_ticket: boolean;
 }
+
+export interface LinkedPretixTelegramEvent {
+  telegramChatID: string | null;
+  eventName: string;
+  configEventID: string;
+  isLinkedToCurrentChat: boolean;
+}
+
+export interface ChatIDWithEventIDs {
+  telegramChatID: string;
+  ticketEventIds: string[];
+}
+export interface UserIDWithChatIDs {
+  telegramUserID: string;
+  telegramChatIDs: string[];
+}
+
+export type ChatIDWithEventsAndMembership = ChatIDWithEventIDs & {
+  isChatMember: boolean;
+};
